+++ src/lib/types/proposal.ts
@@ -76,12 +76,17 @@ export interface ProposalDeposit {
   txHash: string;
 }
 
+interface Message {
+  "@type": string;
+  [key: string]: unknown;
+}
+
 export interface ProposalData extends Proposal {
   createdHeight: Nullable<number>;
   createdTimestamp: Nullable<Date>;
   createdTxHash: Nullable<string>;
   description: string;
-  messages: Nullable<unknown[]>;
+  messages: Nullable<Message[]>;
   metadata: string;
   proposalDeposits: ProposalDeposit[];
   resolvedTimestamp: Nullable<Date>;